pipeline/measurement/pyproject.toml
[tool.poetry]
name = "Labautomation Pipeline"
version = "1.0"
description = "Automation pipeline for chlamydomonas reinhardtii measurements"
authors = ["Cedric K. Brinkmann <cedric.marburg@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.5"
attrs = "^21.4.0"
pandas = "^1.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"